got the webpage to keep track of the username
realized that html is a stateless language so it doesn't keep track of
variables when transitioning from one page to another. Furthermore,
realized that there was an option where you could save a value, but only
for a current session per window, which was perfect because there can
only be one user logged on at one instance in each window.
-able to portray name when logged in.
**took a lot longer than i expected due to the fact that i didn't
realize that html was a stateless language.